moved warden configuration, added created user
