+++ app/models/user.rb
@@ -1,5 +1,5 @@
 class User < ActiveRecord::Base
   has_secure_password
 
-  validates: :username, :password presence: true
+#  validates :email, :password, presence: true
 end
