+++ app/views/register.erb
@@ -5,8 +5,8 @@ Please register!
  <input type="text" name="email" id="email">
  <label for="password">Password</label>
  <input type="text" name="password" id="password">
- <label for="password_again">Password again</label>
- <input type="text" name="password_again" id="password_again">
+ <label for="password_confirmation">Password again</label>
+ <input type="text" name="password_confirmation" id="password_confirmation">
  <br><br>
  <input type="submit" value="Submit">
 </form>
