+++ config/environment.rb
@@ -13,6 +13,7 @@ require 'sinatra/flash'
 require 'sinatra/redirect_with_flash'
 require 'sinatra/reloader' if development?
 require 'pry' unless production?
+require 'warden'
 
 require 'erb'
 
@@ -34,6 +35,50 @@ class SlowFoodApp < Sinatra::Base
   set :views, File.join(APP_ROOT, 'app', 'views')
   set :public_folder, File.join(APP_ROOT, 'public')
   set :show_exceptions, false
+  # Warden configuration code
+  use Rack::Session::Cookie, secret: "IdoNotHaveAnySecret"
+  #use Rack::Flash, accessorize: [:error, :success]
+
+  use Warden::Manager do |manager|
+    manager.default_strategies :password
+    manager.failure_app = SlowFoodApp
+    manager.serialize_into_session {|user| user.id}
+    manager.serialize_from_session {|id| Datastore.for(:user).find_by_id(id)}
+  end
+
+  Warden::Manager.before_failure do |env,opts|
+    env['REQUEST_METHOD'] = 'POST'
+  end
+
+  Warden::Strategies.add(:password) do
+    def valid?
+      params["email"] || params["password"]
+    end
+
+    def authenticate!
+      user = Datastore.for(:user).find_by_email(params["email"])
+      if user && user.authenticate(params["password"])
+        success!(user)
+      else
+        fail!("Could not log in")
+      end
+    end
+  end
+
+  def warden_handler
+    env['warden']
+  end
+
+  def check_authentication
+    unless warden_handler.authenticated?
+      redirect '/login'
+    end
+  end
+
+  def current_user
+    warden_handler.user
+  end
+
 end
 
 Dir[APP_ROOT.join('app', 'controllers', '*.rb')].each { |file| require file }
